# Conflicts:
#	FieldOpt/FieldOpt-WellIndexCalculator
#	FieldOpt/Model/model.cpp
#	FieldOpt/Model/model.h
#	FieldOpt/Model/properties/variable_property_container.cpp
#	FieldOpt/Model/properties/variable_property_container.h
#	FieldOpt/Model/wells/control.cpp
#	FieldOpt/Model/wells/control.h
#	FieldOpt/Model/wells/well.cpp
#	FieldOpt/Model/wells/well.h
#	FieldOpt/Model/wells/wellbore/trajectory.cpp
#	FieldOpt/Model/wells/wellbore/trajectory.h
#	FieldOpt/Model/wells/wellbore/wellspline.cpp
#	FieldOpt/Model/wells/wellbore/wellspline.h
#	FieldOpt/Optimization/CMakeLists.txt
#	FieldOpt/Optimization/case.h
#	FieldOpt/Optimization/case_handler.cpp
#	FieldOpt/Optimization/optimizers/GSS.cpp
#	FieldOpt/Optimization/optimizers/GSS.h
#	FieldOpt/Optimization/optimizers/compass_search.cpp
#	FieldOpt/Runner/logger.cpp
#	FieldOpt/Runner/logger.h
#	FieldOpt/Runner/runners/abstract_runner.cpp
#	FieldOpt/Runner/runners/abstract_runner.h
#	FieldOpt/Runner/runtime_settings.cpp
#	FieldOpt/Settings/CMakeLists.txt
#	FieldOpt/Settings/model.cpp
#	FieldOpt/Settings/model.h
#	FieldOpt/Settings/optimizer.cpp
#	FieldOpt/Settings/optimizer.h
#	FieldOpt/Settings/settings.cpp
#	FieldOpt/Settings/settings.h
#	FieldOpt/Settings/simulator.cpp
#	FieldOpt/Simulation/simulator_interfaces/driver_file_writers/driver_parts/ecl_driver_parts/schedule_section.cpp
#	FieldOpt/Simulation/simulator_interfaces/driver_file_writers/driver_parts/ecl_driver_parts/schedule_section.h
#	FieldOpt/Simulation/simulator_interfaces/driver_file_writers/driver_parts/ecl_driver_parts/wellcontrols.cpp
#	FieldOpt/Simulation/simulator_interfaces/driver_file_writers/driver_parts/ecl_driver_parts/wellcontrols.h
#	FieldOpt/Simulation/simulator_interfaces/driver_file_writers/driver_parts/ecl_driver_parts/welspecs.cpp
#	FieldOpt/Simulation/simulator_interfaces/driver_file_writers/driver_parts/ecl_driver_parts/welspecs.h
#	FieldOpt/Simulation/simulator_interfaces/driver_file_writers/ecldriverfilewriter.cpp
#	FieldOpt/Simulation/simulator_interfaces/eclsimulator.cpp
#	FieldOpt/Simulation/simulator_interfaces/simulator.h
#	FieldOpt/Utilities/debug.hpp

+++ FieldOpt/Simulation/simulator_interfaces/driver_file_writers/driver_parts/adgprs_driver_parts/adgprs_wellcontrols.h
@@ -14,7 +14,8 @@ namespace AdgprsDriverParts {
 class WellControls : public ECLDriverParts::WellControls
 {
 public:
-    WellControls(QList<Model::Wells::Well *> *wells, QList<int> control_times);
+    WellControls(QList<Model::Wells::Well *> *wells,
+                 QList<double> control_times);
 
 
     // DriverPart interface
